[FIXED] Reconnect logic
- When NATS connection is terminally closed (due to MaxReconnects
not being infinite), the bridge is shutdown. The process will now
also exit with a os.Exit(2) to notify docker/system that the exit
was "abnormal".
- On clean shutdown (ctrl+c) the streaming connection should be
closed *before* the NATS connection since with streaming, protocols
are sent to the streaming server and needs the low level NATS
connection.
- The STAN to Kafka connector was not using Errorf() to trace publish
errors to Kafka
- ensureReconnectTimer() was improperly calling itself on error
but continued for NATS/STAN and connectors, which was leading to
the timer be invoked many, many times..

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>