Fix false positive tests (#23)
Currently, the makefile ignores Go test errors because it wants to
always run clean up. This change refactors the makefile to allow Go test
errors to be noticed and ensure that clean up runs.

In addition, it enables GODEBUG=x509ignoreCN=0 during tests to give us
some time to figure out how to generate certs with SANs. This is
temporary and should be removed soon.