Write message indexes correctly in protobuf (#62)
An issue exists in the protobuf serializer where the message indexes
array length and contents are not written correctly causing the
deserialization to fail. This PR correctly sets the array length as well
as the message indexes array contents using varints and also changes the
serializer test to decode the message and check for validity.